Combine LoginRequestAccepted/Denied into a single message type
Changes:
Rename LoginRequestDeniedData to LoginRequestResponseData.
Add IsSuccess bool to LoginRequestResponseData.
Add UnityServer.SendLoginRequestReponse .
Change UnityServer.SendLoginRequestAccepted/Denied to call SendLoginRequestReponse instead.
